requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pandas
pillow
nibabel
scikit-image
tqdm
tensorboard
opencv-python
segment-anything-py